refactor: PostResponse 응답 엔티티 패키지 위치 수정
- com.eventitta.post.dto.request -> com.eventitta.post.dto.response
+ 작성자 닉네임, 프로파일 이미지 URL 필드 추가
